rename: 코드 및 파일 양의 증가에 따른 패키지 구조 조정
