+++ lib/data/models/dtos/request_product_search_dto.dart
