+++ lib/data/models/dtos/response_page_summarized_product_dto.dart
