+++ lib/data/models/dtos/summarized_product_dto.dart
