+++ lib/data/models/els_product.dart
