+++ lib/ui/screens/main_screen.dart
@@ -1,5 +1,5 @@
-import 'package:elswhere/screens/product_screen.dart';
-import 'package:elswhere/widgets/drawer_widget.dart';
+import 'package:elswhere/ui/screens/product_screen.dart';
+import 'package:elswhere/ui/widgets/drawer_widget.dart';
 import 'package:flutter/material.dart';
 import '../widgets/bottom_navigation_bar_widget.dart';
 import 'home_screen.dart';
