+++ lib/ui/views/holding_product_list_view.dart
@@ -1,6 +1,6 @@
+import 'package:elswhere/data/providers/els_products_provider.dart';
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
-import '../providers/els_products_provider.dart';
 import '../widgets/els_product_card.dart';
 
 class HoldingProductListView extends StatelessWidget {
