+++ lib/ui/widgets/explain_card.dart
