+++ lib/ui/widgets/holding_product_card.dart
@@ -1,5 +1,4 @@
-import 'package:elswhere/resources/app_resource.dart';
-import 'package:elswhere/widgets/rounded_text_container.dart';
+import 'package:elswhere/config/app_resource.dart';
 import 'package:flutter/material.dart';
 import 'package:intl/intl.dart';
