+++ lib/ui/widgets/rounded_text_container.dart
@@ -1,6 +1,3 @@
-
-
-import 'package:elswhere/resources/app_resource.dart';
 import 'package:flutter/material.dart';
