+++ lib/ui/widgets/total_outcome_card.dart
@@ -1,7 +1,7 @@
 import 'dart:ui';
 
-import 'package:elswhere/resources/app_resource.dart';
-import 'package:elswhere/widgets/rounded_text_container.dart';
+import 'package:elswhere/config/app_resource.dart';
+import 'package:elswhere/ui/widgets/rounded_text_container.dart';
 import 'package:flutter/material.dart';
 import 'package:intl/intl.dart';
